# RFC: specialize some varargs