@masonprotter

benchmark script
```
@btime (x->last(map(sin, fill(x,100))))'(0.1);
@btime (x->last(map(atan, fill(x,100), fill(2x, 100))))'(0.1);
@btime sin'(0.5);
@btime sin''(0.5);
@btime sin'''(0.5);
```

## Julia 1.4
### Without this PR
```julia
julia> @btime (x->last(map(sin, fill(x,100))))'(0.1);
  17.075 μs (769 allocations: 22.88 KiB)

julia> @btime (x->last(map(atan, fill(x,100), fill(2x, 100))))'(0.1);
  23.213 μs (988 allocations: 37.02 KiB)

julia> @btime sin'(0.5);
  1.346 ns (0 allocations: 0 bytes)

julia> @btime sin''(0.5);
  12.902 ns (0 allocations: 0 bytes)

julia> @btime sin'''(0.5);
  703.346 μs (2664 allocations: 68.30 KiB)
```

### Only with interface.jl changes
```julia
julia> @btime (x->last(map(sin, fill(x,100))))'(0.1);
  15.689 μs (770 allocations: 22.89 KiB)

julia> @btime (x->last(map(atan, fill(x,100), fill(2x, 100))))'(0.1);
  21.278 μs (988 allocations: 37.02 KiB)

julia> @btime sin'(0.5);
  1.247 ns (0 allocations: 0 bytes)

julia> @btime sin''(0.5);
  10.701 ns (0 allocations: 0 bytes)

julia> @btime sin'''(0.5);
  679.070 μs (2626 allocations: 67.69 KiB)
```

### With interface.jl and interface2.jl changes
```julia
julia> @btime (x->last(map(sin, fill(x,100))))'(0.1);
  17.219 μs (770 allocations: 22.89 KiB)

julia> @btime (x->last(map(atan, fill(x,100), fill(2x, 100))))'(0.1);
  22.907 μs (988 allocations: 37.02 KiB)

julia> @btime sin'(0.5);
  1.247 ns (0 allocations: 0 bytes)

julia> @btime sin''(0.5);
  11.414 ns (0 allocations: 0 bytes)

julia> @btime sin'''(0.5);
  674.351 μs (2617 allocations: 67.55 KiB)
```


## Julia 1.5

### Without this PR
```julia
julia> @btime (x->last(map(sin, fill(x,100))))'(0.1);
  14.080 μs (770 allocations: 23.55 KiB)

julia> @btime (x->last(map(atan, fill(x,100), fill(2x, 100))))'(0.1);
  19.151 μs (989 allocations: 37.94 KiB)

julia> @btime sin'(0.5);
  1.246 ns (0 allocations: 0 bytes)

julia> @btime sin''(0.5);
  11.812 ns (0 allocations: 0 bytes)

julia> @btime sin'''(0.5);
  531.809 μs (2410 allocations: 86.27 KiB)
```

### With interface.jl and interface2.jl changes
```julia
julia> @btime (x->last(map(sin, fill(x,100))))'(0.1);
  14.947 μs (770 allocations: 23.55 KiB)

julia> @btime (x->last(map(atan, fill(x,100), fill(2x, 100))))'(0.1);
  19.142 μs (989 allocations: 37.94 KiB)

julia> @btime sin'(0.5);
  1.279 ns (0 allocations: 0 bytes)

julia> @btime sin''(0.5);
  10.987 ns (0 allocations: 0 bytes)

julia> @btime sin'''(0.5);
  509.693 μs (2310 allocations: 83.59 KiB)
```