@MasonProtter 2020-05-18
Thanks for running these. Much less of an improvement that I had hoped, but it could be that there's just different bottlenecks to be tackled. Perhaps something in IRTools.jl itself. 

What branch do you use for "without this PR"? Here's what I see on master:
```julia
julia> begin
       using Zygote
       @btime (x->last(map(sin, fill(x,100))))'(0.1)
       @btime (x->last(map(atan, fill(x,100), fill(2x, 100))))'(0.1)
       @btime sin'(0.5)
       @btime sin''(0.5)
       @btime sin'''(0.5)
       end;
  18.740 μs (769 allocations: 22.88 KiB)
  24.790 μs (988 allocations: 37.02 KiB)
  1.299 ns (0 allocations: 0 bytes)
  9.058 ns (0 allocations: 0 bytes)
  1.317 ms (4652 allocations: 114.19 KiB)

julia> versioninfo()
Julia Version 1.4.1
Commit 381693d3df* (2020-04-14 17:20 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
  CPU: AMD Ryzen 5 2600 Six-Core Processor
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-8.0.1 (ORCJIT, znver1)
Environment:
  JULIA_NUM_THREADS = 6
```
and on your branch:
```julia
julia> begin
       using Zygote
       @btime (x->last(map(sin, fill(x,100))))'(0.1)
       @btime (x->last(map(atan, fill(x,100), fill(2x, 100))))'(0.1)
       @btime sin'(0.5)
       @btime sin''(0.5)
       @btime sin'''(0.5)
       end;
  17.320 μs (770 allocations: 22.89 KiB)
  23.600 μs (988 allocations: 37.02 KiB)
  1.299 ns (0 allocations: 0 bytes)
  9.018 ns (0 allocations: 0 bytes)
  740.627 μs (2617 allocations: 67.67 KiB)
```
so `sin'''` seems way worse for me on master than your "without this PR".

But pursuing SpecializeVarargs.jl just to make higher order derivatives faster seems like putting a bandaid on a headwound. For that, we  either need ways to simplify IR after a derivative so that the IR doesn't combinatorially explode with higher order derivatives, or we need a way to generate less IR bloat in the first place using something like Taylor rules.