@oxinabox 2020-05-19
Interesting, I thought i was using master...

The whole thing of nested Zygote can't be solved via Taylor mode.
Because the only reason you were ever nest Zygote is my mistake.
E.g. using Zygote on come large bit of code from a package that happens to use Zygote internally (and you didn't know about it).
The majority of times you want a second derivative you want to do forward over reverse, or forward over forward. If you do want reverse over reverse i hear Zygote over ReverseDiff works great already.
The `sin'''(x)` code is interesting for benchmarking reasons only.
It is some of the most pathological code one might want to do with AD.
Almost nothing in the real world that someone does on purpose is going to be that hard